Components for messages have been moved to another module.
